+++ tests/Feature/Console/Commands/DatabaseOrchestrateServiceCommandTest.php
@@ -3,10 +3,10 @@
 namespace Payavel\Orchestration\Tests\Feature\Console\Commands;
 
 use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
+use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
 
 class DatabaseOrchestrateServiceCommandTest extends TestOrchestrateServiceCommand
 {
-    use CreatesDatabaseServiceables;
-
-    public $driver = 'database';
+    use CreatesDatabaseServiceables,
+        SetsDatabaseDriver;
 }
