+++ tests/Traits/SetsDatabaseDriver.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Payavel\Orchestration\Tests\Traits;
+
+use Illuminate\Support\Facades\Artisan;
+use Illuminate\Support\Facades\Config;
+
+trait SetsDatabaseDriver
+{
+    /**
+     * Set the driver.
+     *
+     * @return void
+     */
+    protected function setDriver()
+    {
+        Config::set('orchestration.defaults.driver', 'database');
+
+        Artisan::call('vendor:publish', [
+            '--tag' => 'payavel-migrations'
+        ]);
+    }
+}
