+++ tests/Unit/ConfigServiceTest.php
@@ -3,10 +3,10 @@
 namespace Payavel\Orchestration\Tests\Unit;
 
 use Payavel\Orchestration\Tests\Traits\CreatesConfigServiceables;
+use Payavel\Orchestration\Tests\Traits\SetsConfigDriver;
 
 class ConfigServiceTest extends TestService
 {
-    use CreatesConfigServiceables;
-
-    public $driver = 'config';
+    use CreatesConfigServiceables,
+        SetsConfigDriver;
 }
