+++ tests/Unit/DatabaseServiceTest.php
@@ -3,10 +3,10 @@
 namespace Payavel\Orchestration\Tests\Unit;
 
 use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
+use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
 
 class DatabaseServiceTest extends TestService
 {
-    use CreatesDatabaseServiceables;
-
-    public $driver = 'database';
+    use CreatesDatabaseServiceables,
+        SetsDatabaseDriver;
 }
