+++ tests/Unit/FakeConfigServiceTest.php
@@ -3,11 +3,12 @@
 namespace Payavel\Orchestration\Tests\Unit;
 
 use Payavel\Orchestration\Tests\Traits\CreatesConfigServiceables;
+use Payavel\Orchestration\Tests\Traits\SetsConfigDriver;
 
 class FakeConfigServiceTest extends TestService
 {
-    use CreatesConfigServiceables;
+    use CreatesConfigServiceables,
+        SetsConfigDriver;
 
-    public $driver = 'config';
     public $fake = true;
 }
