+++ tests/Unit/FakeDatabaseServiceTest.php
@@ -3,11 +3,12 @@
 namespace Payavel\Orchestration\Tests\Unit;
 
 use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
+use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
 
 class FakeDatabaseServiceTest extends TestService
 {
-    use CreatesDatabaseServiceables;
+    use CreatesDatabaseServiceables,
+        SetsDatabaseDriver;
     
-    public $driver = 'database';
     public $fake = true;
 }
